src/litutils/requirements.txt
pydantic
pydantic-yaml
# torch - install it manually
torchvision
torchaudio
pytorch-lightning
torchmetrics
tensorboard
pandas
optuna
wandb
numpy
seaborn
rich